backend/requirements.txt
fastapi 
sqlalchemy 
pymysql 
python-dotenv 
uvicorn
